Move _createElement from Task components to Card & CardEdit components; refactor components
