+++ src/components/card-edit.js
@@ -1,5 +1,8 @@
 import Task from './task';
+import Text from './task/text';
+import SettingsControls from './task/settings-controls';
 import FormControls from './task/form-controls';
+import {createElement, renderElement} from '../helpers';
 
 export default class CardEdit extends Task {
   constructor(taskData) {
@@ -7,6 +10,8 @@ export default class CardEdit extends Task {
 
     this._init(taskData);
 
+    this._text = new Text(taskData);
+    this._settings = new SettingsControls(taskData);
     this._formControls = new FormControls(taskData);
   }
 
@@ -14,5 +19,19 @@ export default class CardEdit extends Task {
     const form = this.getElement().querySelector(`form`);
 
     form.addEventListener(`submit`, handler);
+  }
+
+  _createElement() {
+    const element = createElement(this._getTmpl());
+    const innerElement = element.querySelector(`.card__inner`);
+
+    renderElement(innerElement, [
+      this._getColorbarElement(),
+      this._text,
+      this._settings,
+      this._formControls
+    ]);
+
+    return element;
   }
 }
