+++ src/components/card.js
@@ -1,12 +1,16 @@
 import Task from './task';
+import Text from './task/text';
+import SettingsText from './task/settings-text';
 import CardControls from './task/card-controls';
+import {createElement, renderElement} from '../helpers';
 
 export default class Card extends Task {
   constructor(taskData) {
     super();
 
     this._init(taskData);
-
+    this._text = new Text(taskData);
+    this._settings = new SettingsText(taskData);
     this._cardControls = new CardControls(taskData);
   }
 
@@ -16,5 +20,19 @@ export default class Card extends Task {
 
   setEditBtnHandler(handler) {
     this._cardControls.setEditBtnHandler(handler);
+  }
+
+  _createElement() {
+    const element = createElement(this._getTmpl());
+    const innerElement = element.querySelector(`.card__inner`);
+
+    renderElement(innerElement, [
+      this._cardControls,
+      this._getColorbarElement(),
+      this._text,
+      this._settings
+    ]);
+
+    return element;
   }
 }
