+++ src/components/task/index.js
@@ -1,7 +1,5 @@
 import AbstractComponent from '../abstract-component';
-import Text from './text';
-import Settings from './settings';
-import {getClass, createElement, renderElement} from '../../helpers';
+import {getClass, createElement} from '../../helpers';
 
 export default class Task extends AbstractComponent {
   constructor() {
@@ -24,9 +22,6 @@ export default class Task extends AbstractComponent {
     this._isRepeat = isRepeat;
     this._isDeadline = isDeadline;
     this._isEdit = isEdit;
-
-    this._text = new Text(taskData);
-    this._settings = new Settings(taskData);
   }
 
   _getMods() {
@@ -73,27 +68,6 @@ export default class Task extends AbstractComponent {
     </div>`;
 
     return createElement(markup);
-  }
-
-  _createElement() {
-    const element = createElement(this._getTmpl());
-    const innerElement = element.querySelector(`.card__inner`);
-
-    if (this._cardControls) {
-      renderElement(innerElement, this._cardControls);
-    }
-
-    renderElement(innerElement, [
-      this._getColorbarElement(),
-      this._text,
-      this._settings
-    ]);
-
-    if (this._formControls) {
-      renderElement(innerElement, this._formControls);
-    }
-
-    return element;
   }
 
   _getTmpl() {
